move glue catalog error into provider mod
